+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Only display authorized message cog buttons
 * Show all roles in Admin user card
 * Remove scoped user_roles when channel removed
+* Support invitation delete
 
 ### Bug Fixes